♻️ refactor : 공통 예외 처리 통합
